Split out drivetrains to allow more implementations
