+++ examples/src/main/java/io/github/frc5024/lib5k/examples/teleop_drive/subsystems/DriveTrain.java
@@ -4,13 +4,13 @@ import com.ctre.phoenix.motorcontrol.NeutralMode;
 
 import edu.wpi.first.wpilibj.geometry.Rotation2d;
 import io.github.frc5024.common_drive.gearing.Gear;
-import io.github.frc5024.lib5k.bases.drivetrain.implementations.TankDriveTrain;
+import io.github.frc5024.lib5k.bases.drivetrain.implementations.OpenLoopTankDriveTrain;
 import io.github.frc5024.lib5k.examples.teleop_drive.RobotConfig;
 import io.github.frc5024.lib5k.hardware.common.sensors.interfaces.CommonEncoder;
 import io.github.frc5024.lib5k.hardware.ctre.motors.ExtendedTalonSRX;
 import io.github.frc5024.lib5k.hardware.kauai.gyroscopes.NavX;
 
-public class DriveTrain extends TankDriveTrain {
+public class DriveTrain extends OpenLoopTankDriveTrain {
     private static DriveTrain instance = null;
 
     // Motors
