+++ lib5k/src/main/java/io/github/frc5024/lib5k/bases/drivetrain/implementations/TankDriveTrain.java
@@ -30,41 +30,19 @@ public abstract class TankDriveTrain extends AbstractDriveTrain {
     // Open loop control
     private DifferentialVoltages openLoopGoal = new DifferentialVoltages();
 
-    // Control loops
-    private Controller distanceController;
-    private Controller rotationController;
-
-    // Timer for actions
-    private Timer actionTimer = new Timer();
-
     // Front side
     private Chassis.Side frontSide = Chassis.Side.kFront;
 
     // Max speed percent
-    private double maxSpeedPercent = 1.0;
+    protected double maxSpeedPercent = 1.0;
 
     // Localization
     private DifferentialDriveOdometry localizer;
 
     /**
-     * Create a new TankDriveTrain that can only be controlled in open-loop mode
+     * Create a new TankDriveTrain
      */
     public TankDriveTrain() {
-        this(null, null);
-    }
-
-    /**
-     * Create a new TankDriveTrain
-     * 
-     * @param distanceController Controller for distance control
-     * @param rotationController Controller for rotation control
-     */
-    public TankDriveTrain(Controller distanceController, Controller rotationController) {
-
-        // Set controllers
-        this.distanceController = distanceController;
-        this.rotationController = rotationController;
-
         // Set up the localizer
         localizer = new DifferentialDriveOdometry(new Rotation2d());
     }
@@ -83,158 +61,6 @@ public abstract class TankDriveTrain extends AbstractDriveTrain {
             logger.log("An Open-Loop goal of NULL was passed to TankDriveTrain. Failed to write to motors!",
                     Level.kWarning);
         }
-    }
-
-    @Override
-    protected void handleAutonomousRotation(StateMetadata<State> meta, Rotation2d goalHeading, Rotation2d epsilon) {
-
-        if (meta.isFirstRun()) {
-
-            // Handle this being an open-loop-only drivetrain
-            if (rotationController == null || distanceController == null) {
-                logger.log(
-                        "This drivetrain does not support autonomous control. Please pass controllers into its constructor",
-                        Level.kWarning);
-                // Switch to open loop control
-                setOpenLoop(new DifferentialVoltages());
-                return;
-            }
-
-            logger.log("Switched to rotation control");
-            logger.log(String.format("Turning to %s with epsilon of %s", goalHeading.toString(), epsilon.toString()));
-
-            // Reset the controller
-            rotationController.reset();
-
-            // Configure the controller
-            rotationController.setEpsilon(epsilon.getDegrees());
-            rotationController.setReference(0);
-
-            // Reset and start the action timer
-            actionTimer.reset();
-            actionTimer.start();
-        }
-
-        // Get the current heading
-        Rotation2d currentHeading = getPose().getRotation();
-
-        // Determine the error from the goal
-        double error = RotationMath.getAngularErrorDegrees(currentHeading, goalHeading);
-
-        // Calculate the needed output
-        double output = rotationController.calculate(error) * RR_HAL.MAXIMUM_BUS_VOLTAGE * maxSpeedPercent;
-
-        // Write voltage outputs
-        handleVoltage(output, -output);
-
-        // Handle reaching the goal heading
-        if (rotationController.atReference()) {
-
-            // Reset the controller
-            rotationController.reset();
-
-            // Stop the motors
-            handleVoltage(0, 0);
-            actionTimer.stop();
-
-            // Switch to open loop control
-            setOpenLoop(new DifferentialVoltages());
-
-            // Log success
-            logger.log(String.format("Reached heading goal after %.2f seconds", actionTimer.get()));
-        }
-
-    }
-
-    @Override
-    protected void handleDrivingToPose(StateMetadata<State> meta, Translation2d goalPose, Translation2d epsilon) {
-
-        if (meta.isFirstRun()) {
-
-            // Handle this being an open-loop-only drivetrain
-            if (rotationController == null || distanceController == null) {
-                logger.log(
-                        "This drivetrain does not support autonomous control. Please pass controllers into its constructor",
-                        Level.kWarning);
-                // Switch to open loop control
-                setOpenLoop(new DifferentialVoltages());
-                return;
-            }
-
-            logger.log("Switched to pose control");
-            logger.log(String.format("Driving to pose: %s", goalPose));
-
-            // Reset the controllers
-            rotationController.reset();
-            distanceController.reset();
-
-            // Configure the controllers
-            rotationController.setEpsilon(1.0);
-            rotationController.setReference(0.0);
-            distanceController.setEpsilon((epsilon.getX() + epsilon.getY()) / 2);
-            distanceController.setReference(0.0);
-
-            // Reset and start the action timer
-            actionTimer.reset();
-            actionTimer.start();
-        }
-
-        // Get the robot's current pose
-        Pose2d currentPose = getPose();
-
-        // Calculate positional error
-        Translation2d error = goalPose.minus(currentPose.getTranslation());
-
-        // Get the hypot to determine scalar distance
-        double distanceError = currentPose.getTranslation().getDistance(goalPose) * -1;
-        // Math.sqrt(Math.pow((goalPose.getX() - currentPose.getTranslation().getX()), 2)
-        //         + Math.pow((goalPose.getY() - currentPose.getTranslation().getY()), 2));
-
-        // Calculate clockwise-positive rotational error
-        Rotation2d angularError = Rotation2d.fromDegrees(Math.toDegrees(Math.atan2(error.getY(), error.getX())) * -1);
-
-        // Calculate speed multiplier based on distance from target.
-        // This lets the robot curve towards the target, instead of snapping to it.
-        // This is a trick I learned from a programmer at 1114. It provides really
-        // smooth outputs
-        // https://bitbucket.org/kaleb_dodd/simbot2019public/src/abc56f5220b5c94bca216f86e3b6b5757d0ffeff/src/main/java/frc/subsystems/Drive.java#lines-337
-        double speedMul = ((-1 * (Math.min(Math.abs(angularError.getDegrees()), 90.0)) / 90.0) + 1);
-
-        // Calculate needed throttle
-        double throttleOutput = distanceController.calculate(distanceError);
-
-        // Restrict throttle output
-        throttleOutput *= speedMul;
-
-        // Calculate rotation PIF
-        double turnOutput = rotationController.calculate(angularError.getDegrees());
-
-        // Calculate motor outputs
-        DifferentialVoltages voltages = DifferentialVoltages.fromThrottleAndSteering(throttleOutput, turnOutput);
-
-        // logger.log(voltages.toString());
-        logger.log(new DifferentialVoltages(throttleOutput, turnOutput).toString());
-
-        // Write output frame
-        handleVoltage(voltages.getLeftVolts(), voltages.getRightVolts());
-
-        // If the robot is at its goal, we are done
-        if (RobotMath.epsilonEquals(currentPose.getTranslation(), goalPose, epsilon)) {
-            // Reset the controllers
-            rotationController.reset();
-            distanceController.reset();
-
-            // Stop the motors
-            handleVoltage(0, 0);
-            actionTimer.stop();
-
-            // Switch to open loop control
-            setOpenLoop(new DifferentialVoltages());
-
-            // Log success
-            logger.log(String.format("Reached pose goal after %.2f seconds", actionTimer.get()));
-        }
-
     }
 
     /**
@@ -386,13 +212,5 @@ public abstract class TankDriveTrain extends AbstractDriveTrain {
         constantCurvatureEnabled = false;
         frontSide = Chassis.Side.kFront;
         maxSpeedPercent = 1.0;
-
-        // Reset controllers
-        if (rotationController != null) {
-            rotationController.reset();
-        }
-        if (distanceController != null) {
-            distanceController.reset();
-        }
     }
 }
